fix(executeATCCheck): correctly escape namespaces
The ATC check failed when the name of a development object included a
namespace. Prior to this change, the generated request body looked like
this:

  ...
  <adtcore:objectReferences>
    <adtcore:objectReference
      adtcore:uri="/sap/bc/adt/ddic/tables//ABC/MY_TABLE/source/main" />
    <adtcore:objectReference
      adtcore:uri="/sap/bc/adt/oo/interfaces//ABC/MY_INTERFACE" />
  </adtcore:objectReferences>
  ...

Due to the unescaped slashes in the namespace, the SAP system failed
with a 400 Bad Request error:

  URI-Mapping cannot be performed due to invalid URI:
  /sap/bc/adt/ddic/tables//ABC/MY_TABLE/source/main

By ensuring the namespaces are properly escaped (e.g. %2FABC%2FMY_TABLE)
for all object types, the ATC check now works as expected. However, for
classes this issue was already fixed in commit 4ff5271c
(gCTSExecuteABAPQualityCheck_namespace_objects (#4623), 2023-10-11).