Add better error logging to isMainPackage
In my project, i try to use the `packages` setting to build multiple binaries at the same time, which triggers the code path leading to `isMainPackage`. Unfortunately, inside all of this, the go binary fails with exit code 1:

```
 info  golangBuild - running command: go list -f {{ .Name }} ./cmd/run1
fatal golangBuild - execution of golang build failed - failed to calculate output binaries or directory, error: running command 'go' failed: cmd.Run() failed: exit status 1
```

It must be exactly the go command inside isMainPackage, because in the loop to get the output paths, no other go binary is executed, and my second package (run2) is not logged.

When I run the command `go list -f "{{ .Name }}" ./cmd/run1` locally, I don't get any error, the command outputs `main` instead -- which is the non failure path.

Unfortunately, when the go binary fails in this case, we don't see any error output, as console output is entirely redirected to `outBuffer`, which is not returned.

This PR improves this by using `fmt.Errorf` to attach the outBuffer to the error. Using `%w` to wrap the original error, any introspection using `errors.Is` and `errors.As` would still work (although noone seems to do this). I also made sure to only add the additional context of the error output _after_ the original error, so the current error message stays unchanged -- there is just more context at the end of it.

Please consider merging this change to improve the error output.